old-tests/submission/Opera/timeelement/001.html

<!doctype html>
<html>
	<head>
		<title>HTML 5 time element API</title>
		<style type="text/css">
#time { visibility: hidden; }
		</style>
		<link rel="help" href="http://www.w3.org/TR/html5/the-time-element.html#the-time-element">
		<script type="text/javascript" src="/resources/testharness.js"></script>
		<script type="text/javascript" src="/resources/testharnessreport.js"></script>
	</head>
	<body>

		<noscript><p>Enable JavaScript and reload</p></noscript>
		<div id="log"></div>
		<!-- intentionally nested to test parsing rules -->
		<p id="time"><time pubdate datetime="2000-02-01T03:04:05Z">Dummy text <time>2001-06-07T<time>08:09<time></time></time>Z</time></time></p>
		<script type="text/javascript">

//Per specification;
//"If the datetime attribute is present, the user agent should convey the attribute's value to the user when rendering the element." #attr-time-datetime
//"time[datetime] { binding: time; }" #the-time-element-0
//This testsuite does not test whether the UA converts the displayed date into a user-preferred format (eg. using XBL2).
//This testsuite does not test whether the UA uses the datetime attribute or textContent when displaying this reformatted date.

function makeTime(pubDate,dateTime,contents,dateTimeProp) {
	var timeEl = document.createElement('time');
	if( pubDate ) {
		timeEl.setAttribute('pubdate','pubdate');
	}
	if( dateTime ) {
		timeEl.setAttribute('datetime',dateTime);
	}
	if( contents ) {
		timeEl.innerHTML = contents;
	}
	if( dateTimeProp ) {
		timeEl.dateTime = dateTimeProp;
	}
	return timeEl;
}

var stamp2000 = 949374245000,   //2000-02-01T03:04:05Z      Generic datetime within 32 bits of Unix epoch
    ymd2000 = 949363200000,     //2000-02-01T00:00:00Z      Generic date within 32 bits of Unix epoch
    ymd10000 = 253402300800000, //10000-02-011T00:00:00Z    Date with 5 digit year, within 63 bits of Unix epoch
    hm1970 = 11040000,          //1970-01-01T03:04:00Z      Time with hours and minutes on first day of 1970
    h24m1970 = 47040000,        //1970-01-01T13:04:00Z      Time after 12:00 with hours and minutes on first day of 1970
    hms1970 = 11045000,         //1970-01-01T03:04:05Z      Time with hours, minutes and seconds on first day of 1970
    hmsf1970 = 11045500,        //1970-01-01T03:04:05.5Z    Time with hours, minutes and x.y fractional seconds on first day of 1970
    hmsfff1970 = 11045555,      //1970-01-01T03:04:05.555Z  Time with hours, minutes and x.yyy fractional seconds on first day of 1970
    round1970 = 59999,          //1970-01-01T00:00:59.999Z  Time with highest number of seconds and milliseconds on first day of 1970
    stampms2000 = 949374245123; //2000-02-01T03:04:05.123Z  Generic datetime with fractional seconds within 32 bits of Unix epoch
var timep = document.getElementById('time');
var times = timep.getElementsByTagName('time');

//TIME elements
test(function () {
	assert_equals( times.length, 4 );
}, 'HTML parsing should locate 4 time elements in this document');
test(function () {
	assert_true( !!window.HTMLTimeElement );
}, 'HTMLTimeElement should be exposed for prototyping');
test(function () {
	assert_true( makeTime() instanceof window.HTMLTimeElement, 'createElement variant' );
	assert_true( times[0] instanceof window.HTMLTimeElement, 'HTML parsing variant' );
}, 'the time elements should be instanceof HTMLTimeElement');

//pubDate
test(function () {
	assert_true(makeTime(true).pubDate);
	assert_false(makeTime().pubDate);
}, 'the pubdate attribute should be reflected by the .pubDate property');
test(function () {
	var timeEl = makeTime();
	timeEl.pubDate = true;
	assert_true(timeEl.pubDate);
}, 'the pubDate property should be read/write');

//dateTime
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z','2001-02-01T03:04:05Z').dateTime, '2000-02-01T03:04:05Z' );
}, 'the datetime attribute should be reflected by the .dateTime property');
test(function () {
	assert_equals( typeof makeTime().dateTime, 'string', 'typeof test' );
	assert_equals( makeTime().dateTime, '', 'value test' );
}, 'the dateTime IDL property should default to an empty string');
test(function () {
	assert_equals( makeTime(false,false,false,'2000-02-01T03:04:05Z').dateTime, '2000-02-01T03:04:05Z' );
}, 'the dateTime property should be read/write');
test(function () {
	assert_equals( makeTime(false,'go fish').dateTime, 'go fish' );
}, 'the datetime attribute should be reflected by the .dateTime property even if it is invalid');
test(function () {
	assert_equals( makeTime(false,false,'2000-02-01T03:04:05Z').dateTime, '' );
}, 'the datetime attribute should not reflect the textContent');

//valueAsDate
test(function () {
	assert_equals( makeTime().valueAsDate, null );
}, 'the valueAsDate property should be null when no date is provided');
test(function () {
	assert_equals( makeTime(false,'go fish').valueAsDate, null );
}, 'the valueAsDate property should be null when date parsing is unsuccessful');
test(function () {
	assert_equals( makeTime(false,false,'go fish').valueAsDate, null );
}, 'the valueAsDate property should be null when textContent date parsing is unsuccessful');
test(function () {
	assert_true( makeTime(false,'2000-02-01T03:04:05Z').valueAsDate instanceof Date );
}, 'the valueAsDate property should be a date object when date parsing is successful');
test(function () {
	var timeEl = makeTime(false,'2000-02-01T03:04:05Z');
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should be be correct when using the datetime attribute');
test(function () {
	var timeEl = makeTime(false,false,'2000-02-01T03:04:05Z');
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should be be correct when using textContent');
test(function () {
	var timeEl = makeTime(false,false,'2000-02-01T<time>03:04:05<time></time></time>Z');
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should be be correct when using textContent with nested elements');
test(function () {
	assert_equals( makeTime(false,'go fish','2000-02-01T03:04:05Z').valueAsDate, null );
}, 'the valueAsDate property should use the datetime attribute instead of textContent even if datetime fails to parse');
test(function () {
	var timeEl = makeTime( false, false, false, '2000-02-01T03:04:05Z' );
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should update from null when writing to the dateTime property');
test(function () {
	var timeEl = makeTime( false, false, '2001-02-01T03:04:05Z', '2000-02-01T03:04:05Z' );
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should update from an existing value when writing to the dateTime property');
test(function () {
	var timeEl = makeTime( false, false, '2001-02-01T03:04:05Z' );
	//retrieve the property once, to make sure the browser fails if it incorrectly caches it after first access
	//assign to a variable just to catch browsers that incorrectly optimise away the statement without running the getter (Opera's older script engine)
	var ignoreme = timeEl.valueAsDate;
	timeEl.setAttribute('datetime','2000-02-01T03:04:05Z');
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should update from an existing value when writing to the datetime attribute');
test(function () {
	var timeEl = makeTime( false, false, '2001-02-01T03:04:05Z' );
	var ignoreme = timeEl.valueAsDate;
	timeEl.innerHTML = '2000-02-01T03:04:05Z';
	assert_true( timeEl.valueAsDate != null, 'valueAsDate was unexpectedly null' );
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000 );
}, 'the valueAsDate property should update from an existing value when writing to the textContent');
test(function () {
	var timeEl = makeTime(false,'2000-02-01T03:04:05Z');
	timeEl.valueAsDate.tempProperty = true;
	assert_equals( timeEl.valueAsDate.tempProperty, window.undefined );
}, 'the valueAsDate property should retrieve a new Date object every time it is accessed');
test(function () {
	var timeEl = makeTime( false, '2000-02-01T03:04:05Z' );
	try {
		timeEl.valueAsDate = new Date(2001);
	} catch(e) {}
	assert_equals( timeEl.valueAsDate.getTime(), stamp2000, 'date object changed' );
	assert_equals( timeEl.dateTime, '2000-02-01T03:04:05Z', 'dateTime changed' );
}, 'the valueAsDate property should be read-only');
test(function () {
	//WebIDL and ECMAScript 5 - a readonly property has a getter but not a setter
	//ES5 makes [[Put]] fail but not throw
	var failed = false, timeEl = makeTime( false, '2000-02-01T03:04:05Z' );
	try {
		timeEl.valueAsDate = new Date(2001);
	} catch(e) {
		failed = e;
	}
	assert_false(failed);
}, 'writing to valueAsDate should be silently ignored and not throw an error');

//"needs a date" is a document requirement, not a UA requirement
test(function () {
	assert_true( makeTime(true,'03:04:05').valueAsDate != null, 'datetime attribute' );
	assert_true( makeTime(true,false,'03:04:05').valueAsDate != null, 'textContent' );
}, 'datetime attribute should still accept a time string even when the element needs a date');

//"in content" vs not "in content"
test(function () {
	assert_equals( makeTime(false,'\r2000-02-01T03:04:05Z').valueAsDate, null, 'leading \\r' );
	assert_equals( makeTime(false,'\n2000-02-01T03:04:05Z').valueAsDate, null, 'leading \\n' );
	assert_equals( makeTime(false,'\t2000-02-01T03:04:05Z').valueAsDate, null, 'leading \\t' );
	assert_equals( makeTime(false,'\f2000-02-01T03:04:05Z').valueAsDate, null, 'leading \\f' );
	assert_equals( makeTime(false,' 2000-02-01T03:04:05Z').valueAsDate, null, 'leading space' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z\r').valueAsDate, null, 'trailing \\r' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z\n').valueAsDate, null, 'trailing \\n' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z\t').valueAsDate, null, 'trailing \\t' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z\f').valueAsDate, null, 'trailing \\f' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z ').valueAsDate, null, 'trailing space' );
	assert_equals( makeTime( false, false, false, '\r2000-02-01T03:04:05Z' ).valueAsDate, null, 'leading \\r with .dateTime property' );
	assert_equals( makeTime( false, false, false, '\r2000-02-01T03:04:05Z' ).valueAsDate, null, 'leading \\n with .dateTime property' );
	assert_equals( makeTime( false, false, false, '\t2000-02-01T03:04:05Z' ).valueAsDate, null, 'leading \\t with .dateTime property' );
	assert_equals( makeTime( false, false, false, '\f2000-02-01T03:04:05Z' ).valueAsDate, null, 'leading \\f with .dateTime property' );
	assert_equals( makeTime( false, false, false, ' 2000-02-01T03:04:05Z' ).valueAsDate, null, 'leading space with .dateTime property' );
	assert_equals( makeTime( false, false, false, '2000-02-01T03:04:05Z\r' ).valueAsDate, null, 'trailing \\r with .dateTime property' );
	assert_equals( makeTime( false, false, false, '2000-02-01T03:04:05Z\n' ).valueAsDate, null, 'trailing \\n with .dateTime property' );
	assert_equals( makeTime( false, false, false, '2000-02-01T03:04:05Z\t' ).valueAsDate, null, 'trailing \\t with .dateTime property' );
	assert_equals( makeTime( false, false, false, '2000-02-01T03:04:05Z\f' ).valueAsDate, null, 'trailing \\f with .dateTime property' );
	assert_equals( makeTime( false, false, false, '2000-02-01T03:04:05Z ' ).valueAsDate, null, 'trailing space with .dateTime property' );
}, 'datetime attribute should not allow leading/trailing whitespace');
test(function () {
	assert_true( makeTime(false,false,'\r\n\t\f 2000-02-01T03:04:05Z\r\n\t\f ').valueAsDate != null );
}, 'textContent should allow leading/trailing whitespace');
test(function () {
	assert_equals( makeTime(false,'A2000-02-01').valueAsDate, null, 'leading A with datetime' );
	assert_equals( makeTime(false,'.2000-02-01').valueAsDate, null, 'leading . with datetime' );
	assert_equals( makeTime(false,'2000-02-01A').valueAsDate, null, 'trailing A with datetime' );
	assert_equals( makeTime(false,'2000-02-01.').valueAsDate, null, 'trailing . with datetime' );
	assert_equals( makeTime(false,'2000-02-01-').valueAsDate, null, 'trailing - with datetime' );
	assert_equals( makeTime(false,false,'A2000-02-01').valueAsDate, null, 'leading A with textContent' );
	assert_equals( makeTime(false,false,'.2000-02-01').valueAsDate, null, 'leading . with textContent' );
	assert_equals( makeTime(false,false,'2000-02-01A').valueAsDate, null, 'trailing A with textContent' );
	assert_equals( makeTime(false,false,'2000-02-01.').valueAsDate, null, 'trailing . with textContent' );
	assert_equals( makeTime(false,false,'2000-02-01-').valueAsDate, null, 'trailing - with textContent' );
}, 'date should not allow leading/trailing characters');

//date formats
test(function () {
	assert_equals( makeTime(false,'2000').valueAsDate, null );
}, 'date parsing should not work with only a specified year');
test(function () {
	assert_equals( makeTime(false,'2000-02').valueAsDate, null );
}, 'date parsing should not work with only a specified year and month');
test(function () {
	assert_true( makeTime(false,'2000-02-01').valueAsDate != null, 'date within 32 bits of Unix epoch' );
	assert_true( makeTime(false,'0001-02-01').valueAsDate != null, '0001 was not accepted as a valid year' );
}, 'date parsing should work with YYYY-MM-DD format');
test(function () {
	assert_equals( makeTime(false,'2000-02-01').valueAsDate.getTime(), ymd2000 );
}, 'date parsing should assume 00:00Z for time');
test(function () {
	assert_equals( makeTime(false,'10000-02-01').valueAsDate.getTime(), ymd10000 );
}, 'date parsing should work with YYYYY-MM-DD format');
//out-of-range dates
	test(function () {
		assert_equals( makeTime(false,'0000-02-01').valueAsDate, null );
	}, 'date parsing should not work with year 0000');
	test(function () {
		assert_equals( makeTime(false,'-0001-02-01').valueAsDate, null );
	}, 'date parsing should not work with year -0001');
	test(function () {
		assert_equals( makeTime(false,'2000-00-01').valueAsDate, null );
	}, 'date parsing should not work with month < 1');
	test(function () {
		assert_equals( makeTime(false,'2000-13-01').valueAsDate, null );
	}, 'date parsing should not work with month > 12');
	test(function () {
		assert_equals( makeTime(false,'2000-1-01').valueAsDate, null );
	}, 'date parsing should not work without a leading zero on months 1-9');
	test(function () {
		assert_equals( makeTime(false,'2000-02-00').valueAsDate, null );
	}, 'date parsing should not work with day < 1');
	test(function () {
		assert_equals( makeTime(false,'2000-01-32').valueAsDate, null, '32nd January' );
		assert_equals( makeTime(false,'2000-02-30').valueAsDate, null, '30th February on a leap year' );
		assert_equals( makeTime(false,'1900-02-29').valueAsDate, null, '29th February on a non-leap century' );
		assert_equals( makeTime(false,'2001-02-29').valueAsDate, null, '29th February on a non-leap year' );
		assert_equals( makeTime(false,'2000-03-32').valueAsDate, null, '32nd March' );
		assert_equals( makeTime(false,'2000-04-31').valueAsDate, null, '31st April' );
		assert_equals( makeTime(false,'2000-05-32').valueAsDate, null, '32nd May' );
		assert_equals( makeTime(false,'2000-06-31').valueAsDate, null, '31st June' );
		assert_equals( makeTime(false,'2000-07-32').valueAsDate, null, '32nd July' );
		assert_equals( makeTime(false,'2000-08-32').valueAsDate, null, '32nd August' );
		assert_equals( makeTime(false,'2000-09-31').valueAsDate, null, '31st September' );
		assert_equals( makeTime(false,'2000-10-32').valueAsDate, null, '32nd October' );
		assert_equals( makeTime(false,'2000-11-31').valueAsDate, null, '31st November' );
		assert_equals( makeTime(false,'2000-12-32').valueAsDate, null, '32nd December' );
	}, 'date parsing should not work with more days than there are in the given month');
	test(function () {
		assert_true( makeTime(false,'2000-02-29').valueAsDate != null, '29th February on a normal leap year' );
	}, 'date parsing should work with leap days');
	test(function () {
		assert_true( makeTime(false,'2001-01-31').valueAsDate != null, '31st January' );
		assert_true( makeTime(false,'2001-02-28').valueAsDate != null, '28th February on a non-leap year' );
		assert_true( makeTime(false,'2001-03-31').valueAsDate != null, '31st March' );
		assert_true( makeTime(false,'2001-04-30').valueAsDate != null, '30th April' );
		assert_true( makeTime(false,'2001-05-31').valueAsDate != null, '31st May' );
		assert_true( makeTime(false,'2001-06-30').valueAsDate != null, '30th June' );
		assert_true( makeTime(false,'2001-07-31').valueAsDate != null, '31st July' );
		assert_true( makeTime(false,'2001-08-31').valueAsDate != null, '31st August' );
		assert_true( makeTime(false,'2001-09-30').valueAsDate != null, '30th September' );
		assert_true( makeTime(false,'2001-10-31').valueAsDate != null, '31st October' );
		assert_true( makeTime(false,'2001-11-30').valueAsDate != null, '30th November' );
		assert_true( makeTime(false,'2001-12-31').valueAsDate != null, '31st December' );
	}, 'date parsing should work with the last day of every non-leap month');
	test(function () {
		assert_equals( makeTime(false,'2000-01-1').valueAsDate, null );
	}, 'date parsing should not work without a leading zero on days 1-9');
	test(function () {
		assert_equals( makeTime(false,'2000-01-01-01').valueAsDate, null );
	}, 'date parsing should not work with extra -xx after the day');
test(function () {
	assert_equals( makeTime(false,'2000-01-01Z').valueAsDate, null, 'timezone Z' );
	assert_equals( makeTime(false,'2000-01-01+01:00').valueAsDate, null, 'timezone +01:00' );
	assert_equals( makeTime(false,'2000-01-01-01:00').valueAsDate, null, 'timezone -01:00' );
}, 'date parsing should not work with timezone');
//making sure browsers do not attempt to understand locale formats
	test(function () {
		assert_equals( makeTime(false,'99-02-01').valueAsDate, null );
	}, 'date parsing should not work with yy-mm-dd format');
	test(function () {
		assert_equals( makeTime(false,'2000/02/01').valueAsDate, null );
	}, 'date parsing should work with yyyy/mm/dd format');
	test(function () {
		assert_equals( makeTime(false,'20-02-2000').valueAsDate, null );
	}, 'date parsing should not work with dd-mm-yyyy format');
	test(function () {
		assert_equals( makeTime(false,'02-20-2000').valueAsDate, null );
	}, 'date parsing should not work with mm-dd-yyyy format');
	test(function () {
		assert_equals( makeTime(false,'20/02/2000').valueAsDate, null );
	}, 'date parsing should not work with dd/mm/yyyy format');
	test(function () {
		assert_equals( makeTime(false,'02/20/2000').valueAsDate, null );
	}, 'date parsing should not work with mm/dd/yyyy format');
	test(function () {
		assert_equals( makeTime(false,'01 Jan 2001').valueAsDate, null );
	}, 'date parsing should not work with dd Mon yyyy format');
	test(function () {
		assert_equals( makeTime(false,'01 January 2001').valueAsDate, null );
	}, 'date parsing should not work with dd Month yyyy format');
	test(function () {
		assert_equals( makeTime(false,'Jan 01 2001').valueAsDate, null );
	}, 'date parsing should not work with Mon dd yyyy format');
	test(function () {
		assert_equals( makeTime(false,'January 01 2001').valueAsDate, null );
	}, 'date parsing should not work with Month dd yyyy format');
	test(function () {
		assert_equals( makeTime(false,'Mon, 01 Jan 2001 00:00:00 GMT').valueAsDate, null, 'variant 1' );
		assert_equals( makeTime(false,'Sat Jan 01 10000 00:00:00 GMT+0000').valueAsDate, null, 'variant 2' );
	}, 'date parsing should not work with long date format');
	test(function () {
		assert_equals( makeTime(false,'949363200000').valueAsDate, null );
	}, 'date parsing should not work with timestamps');

//time format
test(function () {
	assert_equals( makeTime(false,'03').valueAsDate, null );
}, 'time parsing should not work with only a specified hour');
test(function () {
	assert_true( makeTime(false,'03:04').valueAsDate != null );
}, 'time parsing should work with specified hour and minutes');
test(function () {
	assert_equals( makeTime(false,'03:04').valueAsDate.getTime(), hm1970 );
}, 'time parsing should assume 1970-01-01 UTC');
test(function () {
	assert_equals( makeTime(false,'13:04').valueAsDate.getTime(), h24m1970 );
}, 'time parsing should use 24 hour clock');
test(function () {
	assert_equals( makeTime(false,'3:04').valueAsDate, null );
}, 'time parsing should not allow single digit hour');
test(function () {
	assert_equals( makeTime(false,'-3:04').valueAsDate, null );
	assert_equals( makeTime(false,'-03:04').valueAsDate, null );
}, 'time parsing should not allow negative hour');
test(function () {
	assert_equals( makeTime(false,'03:-4').valueAsDate, null );
	assert_equals( makeTime(false,'03:-04').valueAsDate, null );
}, 'time parsing should not allow negative minutes');
test(function () {
	assert_equals( makeTime(false,'03:04.55').valueAsDate, null );
	assert_equals( makeTime(false,'03:04.').valueAsDate, null );
}, 'time parsing should not allow a decimal point after minutes');
test(function () {
	assert_equals( makeTime(false,'03:04:05').valueAsDate.getTime(), hms1970 );
}, 'time parsing should work with specified seconds');
test(function () {
	assert_equals( makeTime(false,'03:04:05.').valueAsDate, null );
}, 'time parsing should not work with just a decimal point after seconds');
test(function () {
	assert_equals( makeTime(false,'03:04:05.5').valueAsDate.getTime(), hmsf1970 );
	assert_equals( makeTime(false,'03:04:05.555').valueAsDate.getTime(), hmsfff1970 );
}, 'time parsing should work with specified fractional seconds');
test(function () {
	assert_equals( makeTime(false,'03:04:05.5551').valueAsDate.getTime(), hmsfff1970 );
}, 'time parsing should work with fractional parts smaller than 1ms');
test(function () {
	assert_equals( makeTime(false,'03:04:05:06').valueAsDate, null );
}, 'time parsing should not allow a colon instead of a decimal point');
test(function () {
	assert_equals( makeTime(false,'03:04:055.5').valueAsDate, null );
}, 'time parsing should not allow seconds to have more than 2 characters before the decimal point');
test(function () {
	assert_equals( makeTime(false,'03:04:05.5.5').valueAsDate, null );
	assert_equals( makeTime(false,'03:04:05.55.').valueAsDate, null );
}, 'time parsing should not allow multiple decimal points');
test(function () {
	assert_equals( makeTime(false,'03:04:5').valueAsDate, null );
	assert_equals( makeTime(false,'03:04:5.5').valueAsDate, null );
}, 'time parsing should not allow single digit seconds');
test(function () {
	assert_equals( makeTime(false,'03:04:-5').valueAsDate, null );
	assert_equals( makeTime(false,'03:04:-05').valueAsDate, null );
}, 'time parsing should not allow negative seconds');
test(function () {
	assert_equals( makeTime(false,'24:00').valueAsDate, null );
}, 'time parsing should not allow hour > 23');
test(function () {
	assert_equals( makeTime(false,'00:60').valueAsDate, null );
}, 'time parsing should not allow minute > 59');
test(function () {
	assert_equals( makeTime(false,'00:00:60').valueAsDate, null );
}, 'time parsing should not allow seconds > 59.9[...]');
test(function () {
	assert_equals( makeTime(false,'00:00:59.9999').valueAsDate.getTime(), round1970 );
}, 'time parsing should floor 59.9999 seconds to the nearest ms to avoid tipping over to 60');
test(function () {
	assert_equals( makeTime(false,'03:04Z').valueAsDate, null, 'timezone Z' );
	assert_equals( makeTime(false,'03:04+01:00').valueAsDate, null, 'timezone +01:00' );
	assert_equals( makeTime(false,'03:04-01:00').valueAsDate, null, 'timezone -01:00' );
}, 'time parsing should not work with timezone');

//datetime format
test(function () {
	assert_equals( makeTime(false,'2000-02-01T 03:04:05Z').valueAsDate, null, 'space after T' );
	assert_equals( makeTime(false,'2000-02-01T 03:04:05Z').valueAsDate, null, 'space before T' );
}, 'datetime parsing should not work with a space in the middle of a datetime string');
test(function () {
	assert_equals( makeTime(false,'2000-02-01 03:04:05Z').valueAsDate, null );
}, 'datetime parsing should not work with a space instead of time separator in a datetime string');
test(function () {
	assert_equals( makeTime(false,'2000-02-01P03:04:05Z').valueAsDate, null );
}, 'datetime parsing should not work with the wrong character as a time separator in a datetime string');
test(function () {
	assert_equals( makeTime(false,'2000-02-01t03:04:05Z').valueAsDate, null );
}, 'the time separator should only accept an upper case T');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T').valueAsDate, null );
}, 'datetime parsing should require time after the separator');
test(function () {
	assert_equals( makeTime(false,'2000-02-01TZ').valueAsDate, null );
}, 'datetime parsing should require time between separator and timezone');
test(function () {
	assert_equals( makeTime(false,'2000-02-01TZ').valueAsDate, null );
}, 'datetime parsing should require time between separator and timezone');
//specific formats for date and time were tested earlier
//only checking basics here in case the UA allows one or the other to be invalid without returning null
test(function () {
	assert_equals( makeTime(false,'2000-02-01T24:00Z').valueAsDate, null, 'out-of-bounds time' );
	assert_equals( makeTime(false,'2000-02-01Taa:bbZ').valueAsDate, null, 'junk in time' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05:00Z').valueAsDate, null, 'junk after time' );
}, 'datetime parsing should require valid time');
test(function () {
	assert_equals( makeTime(false,'2000-02-50T04:04Z').valueAsDate, null, 'out-of-bounds date' );
	assert_equals( makeTime(false,'2000-aa-01T03:04Z').valueAsDate, null, 'junk in date' );
	assert_equals( makeTime(false,'2000-02-01-01T03:04').valueAsDate, null, 'junk after date' );
}, 'datetime parsing should require valid date');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04').valueAsDate, null, 'without seconds' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05').valueAsDate, null, 'with seconds' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05.00').valueAsDate, null, 'with second fragment' );
}, 'datetime parsing should require timezone');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05Z').valueAsDate.getTime(), stamp2000, 'Z' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05.123Z').valueAsDate.getTime(), stampms2000, 'Z with fractional seconds' );
}, 'the timezone Z should act as UTC');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05z').valueAsDate, null );
}, 'the timezone letter should only accept an upper case Z');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05+00:00').valueAsDate.getTime(), stamp2000, '+00:00' );
	assert_equals( makeTime(false,'2000-02-01T04:04:05+01:00').valueAsDate.getTime(), stamp2000, '+01:00' );
	assert_equals( makeTime(false,'2000-02-02T02:04:05+23:00').valueAsDate.getTime(), stamp2000, '+23:00' );
}, 'numeric timezone should offset the parsed time by the given positive hours');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:05:05+00:01').valueAsDate.getTime(), stamp2000, '+00:01' );
	assert_equals( makeTime(false,'2000-02-01T04:03:05+00:59').valueAsDate.getTime(), stamp2000, '+00:59' );
}, 'numeric timezone should offset the parsed time by the given positive minutes');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T04:05:05+01:01').valueAsDate.getTime(), stamp2000, '+01:01' );
	assert_equals( makeTime(false,'2000-02-02T03:03:05+23:59').valueAsDate.getTime(), stamp2000, '+23:59' );
	assert_equals( makeTime(false,'2000-02-02T03:03:05.123+23:59').valueAsDate.getTime(), stampms2000, '+23:59 with fractional seconds' );
}, 'numeric timezone should offset the parsed time by the given positive hours and minutes');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05-00:00').valueAsDate.getTime(), stamp2000, '-00:00' );
	assert_equals( makeTime(false,'2000-02-01T02:04:05-01:00').valueAsDate.getTime(), stamp2000, '-01:00' );
	assert_equals( makeTime(false,'2000-01-31T04:04:05-23:00').valueAsDate.getTime(), stamp2000, '-23:00' );
}, 'numeric timezone should offset the parsed time by the given negative hours');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:03:05-00:01').valueAsDate.getTime(), stamp2000, '-00:01' );
	assert_equals( makeTime(false,'2000-02-01T02:05:05-00:59').valueAsDate.getTime(), stamp2000, '-00:59' );
}, 'numeric timezone should offset the parsed time by the given negative minutes');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T02:03:05-01:01').valueAsDate.getTime(), stamp2000, '-01:01' );
	assert_equals( makeTime(false,'2000-01-31T03:05:05-23:59').valueAsDate.getTime(), stamp2000, '-23:59' );
}, 'numeric timezone should offset the parsed time by the given negative hours and minutes');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05+24:00').valueAsDate, null );
}, 'numeric timezone should not allow hour > 23');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05-24:00').valueAsDate, null );
}, 'numeric timezone should not allow hour < -23');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05+00:60').valueAsDate, null );
}, 'numeric timezone should not allow minutes > 59');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05-00:60').valueAsDate, null );
}, 'numeric timezone should not allow minutes < -59');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05+').valueAsDate, null, '+' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05-').valueAsDate, null, '-' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+10').valueAsDate, null, '+10' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05-10').valueAsDate, null, '-10' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+1:00').valueAsDate, null, '+1:00' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+10:0').valueAsDate, null, '+10:0' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+010:00').valueAsDate, null, '+010:00' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+10:000').valueAsDate, null, '+10:000' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+1000').valueAsDate, null, '+1000' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05-1000').valueAsDate, null, '-1000' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+10:-5').valueAsDate, null, '+10:-5' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05+10:-05').valueAsDate, null, '+10:-05' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05-10:+5').valueAsDate, null, '-10:+5' );
	assert_equals( makeTime(false,'2000-02-01T03:04:05-10:+05').valueAsDate, null, '-10:+05' );
	assert_equals( makeTime(false,'2000-02-01T03:04:0510:-5').valueAsDate, null, '10:-5' );
	assert_equals( makeTime(false,'2000-02-01T03:04:0510:-05').valueAsDate, null, '10:-05' );
	assert_equals( makeTime(false,'2000-02-01T03:04:0510:00').valueAsDate, null, '10:00' );
	assert_equals( makeTime(false,'2000-02-01T03:04:051000').valueAsDate, null, '1000' );
}, 'datetime parsing should not work with invalid numeric timezone formats');
test(function () {
	assert_equals( makeTime(false,'2000-02-01T03:04:05ZA').valueAsDate, null );
	assert_equals( makeTime(false,'2000-02-01T03:04:05GMT').valueAsDate, null );
	assert_equals( makeTime(false,'2000-02-01T03:04:05UTC').valueAsDate, null );
}, 'datetime parsing should not work with non-Z and non-numeric timezones');

		</script>

	</body>
</html>
